Polish BookController responses

+++ src/main/java/com/bookworm/controllers/BookController.java
@@ -4,10 +4,12 @@ import com.bookworm.models.*;
 import com.bookworm.repositories.BookRepository;
 import com.bookworm.repositories.PurchaseRepository;
 import com.bookworm.security.AuthenticatedUser;
+import javax.annotation.Resource;
 import validate.ValidationError;
 import validate.ValidationErrorBuilder;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
 import org.springframework.hateoas.Link;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.linkTo;
 import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
@@ -26,11 +28,13 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+@Scope("singleton")
 @RestController
 @RequestMapping(value = "/books")
 public class BookController {
 
-    protected final HttpServletRequest request; 
+    @Resource
+    private HttpServletRequest request;
     
     @Autowired
     BookRepository bookRepository;
@@ -38,16 +42,18 @@ public class BookController {
     @Autowired
     PurchaseRepository purchaseRepository;
     
-    public BookController(HttpServletRequest request) {
-        this.request = request;
+    public BookController() {
+        
     }
 
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<Book> getBooks() {
-        HttpHeaders headers = new HttpHeaders();
         Iterable<Book> books = bookRepository.findAll();
+        ResponseEntity<Book> response = new ResponseEntity(HttpStatus.NOT_FOUND);
         
         if (books != null) {
+            response = new ResponseEntity(books, HttpStatus.OK);
+            
             for (Book book : books) {
                 Link selfLink = linkTo(BookController.class).slash(book.getBookId()).withSelfRel();
                 Link buyLink = linkTo(methodOn(BookController.class).buyBook(book.getBookId(), 1)).withRel("buyBook");
@@ -56,38 +62,38 @@ public class BookController {
             }
         }
         
-        return new ResponseEntity(books, headers, HttpStatus.OK);
+        return response;
     }
 
     @RequestMapping(value = "/{bookId}", method = RequestMethod.GET)
     public ResponseEntity<Book> getBook(@PathVariable long bookId) {
-        HttpHeaders headers = new HttpHeaders();
         Book findThisBook = bookRepository.findOne(bookId);
+        ResponseEntity<Book> response = new ResponseEntity(HttpStatus.NOT_FOUND);
         
         if (findThisBook != null) {
+            response = new ResponseEntity(findThisBook, HttpStatus.OK);
+            
             Link selfLink = linkTo(BookController.class).slash(findThisBook.getBookId()).withSelfRel();
             Link buyLink = linkTo(methodOn(BookController.class).buyBook(findThisBook.getBookId(), 1)).withRel("buyBook");
             findThisBook.add(selfLink);
             findThisBook.add(buyLink);
-        } else {
-            findThisBook = new Book();
-            Link allLink = linkTo(methodOn(BookController.class).getBooks()).withRel("allBooks");
-            findThisBook.add(allLink);
         }
         
-        return new ResponseEntity(findThisBook, headers, HttpStatus.OK);
+        return response;
     }
     
-    @RequestMapping(method = RequestMethod.POST)
+    @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Book> saveBook(@Validated @RequestBody Book c) {
         HttpHeaders headers = new HttpHeaders();
         Book saveThisBook = bookRepository.save(c);
         
         Link selfLink = linkTo(BookController.class).slash(saveThisBook.getBookId()).withSelfRel();
-        Link allLink = linkTo(methodOn(BookController.class).getBooks()).withRel("allBooks");
+        Link allBooks = linkTo(methodOn(BookController.class).getBooks()).withRel("allBooks");
         saveThisBook.add(selfLink);
-        saveThisBook.add(allLink);
+        saveThisBook.add(allBooks);
+        
         headers.setLocation(linkTo(BookController.class).slash(saveThisBook.getBookId()).toUri());
+        
         return new ResponseEntity(saveThisBook, headers, HttpStatus.CREATED);
     }
     
@@ -96,25 +102,47 @@ public class BookController {
     public ResponseEntity<Purchase> buyBook(@PathVariable long bookId, @RequestBody int amount) {
         AuthenticatedUser member = AuthenticatedUser.fromRequest(request);
         HttpHeaders headers = new HttpHeaders();
+        ResponseEntity<Purchase> response = new ResponseEntity(HttpStatus.NOT_FOUND);
         Book buyThisBook = bookRepository.findOne(bookId);
         Purchase newPurchase = new Purchase();
         
         if (buyThisBook != null) {
-            if (buyThisBook.getStock() - amount >= 0) {
-                bookRepository.reduceStock(buyThisBook.getBookId(), amount);
-                newPurchase = new Purchase(member.id, buyThisBook.getBookId(), amount);
-                purchaseRepository.save(newPurchase);
-            }
+            newPurchase = new Purchase(member.id, buyThisBook.getBookId(), amount);
+            headers.setLocation(linkTo(PurchaseController.class).slash(newPurchase.getPurchaseId()).toUri());
+            response = new ResponseEntity(newPurchase, headers, HttpStatus.CONFLICT);
             
-            headers.setLocation(linkTo(BookController.class).slash(buyThisBook.getBookId()).toUri());
+            if (buyThisBook.getStock() - amount >= 0) {
+                response = new ResponseEntity(newPurchase, headers, HttpStatus.CREATED);
+                bookRepository.reduceStock(buyThisBook.getBookId(), newPurchase.getAmount());
+                purchaseRepository.save(newPurchase);
+                Link selfLink = linkTo(methodOn(PurchaseController.class).getPurchase(newPurchase.getPurchaseId())).withSelfRel();
+                Link allPurchases = linkTo(methodOn(PurchaseController.class).getPurchases()).withRel("allPurchases");
+                Link bookLink = linkTo(methodOn(BookController.class).getBook(newPurchase.getBookId())).withRel("findBook");
+                newPurchase.add(selfLink);
+                newPurchase.add(allPurchases);
+                newPurchase.add(bookLink);
+            }
         }
         
-        return new ResponseEntity(newPurchase, headers, HttpStatus.OK);
+        return response;
     }
 
     @RequestMapping(value = "/{bookId}", method = RequestMethod.DELETE)
-    public void deleteBook(@PathVariable long bookId) {
-        bookRepository.delete(bookId);
+    public ResponseEntity<Book> deleteBook(@PathVariable long bookId) {
+        HttpHeaders headers = new HttpHeaders();
+        Book deleteThisBook = bookRepository.findOne(bookId);
+        ResponseEntity<Book> response = new ResponseEntity(HttpStatus.NOT_FOUND);
+        
+        if (deleteThisBook != null) {
+            headers.setLocation(linkTo(BookController.class).slash(deleteThisBook.getBookId()).toUri());
+            response = new ResponseEntity(deleteThisBook, headers, HttpStatus.OK);
+            
+            bookRepository.delete(bookId);
+            Link allBooks = linkTo(methodOn(BookController.class).getBooks()).withRel("allBooks");
+            deleteThisBook.add(allBooks);
+        }
+                
+        return response;
     }
     
     @ExceptionHandler
